🩹✨ fix(patch): importable tailwindcss; fix @roots/sage error logs (#1703)
- moves the tailwind config handlers out of `@roots/sage` into `@roots/bud-tailwindcss`
- generates virtual modules for tailwind config values aliased to `@tailwind`

Generating the imports can be memory intensive and increase build times, so it is opt-in.

Generating an import for every `theme` key:

```ts
app.tailwind.generateImports()
```

Generating an import for specific keys:

```ts
app.tailwind.generateImports([`colors`, `fontFamily`])
```

Generating an import for specific used keys is a lot less intensive than generating imports for everything. In the future we can add a NormalModule plugin to do this automatically.

Anyway, on to an example of using the generated imports in app code:

```ts
import {black} from '@tailwind/colors'
import {sans} from '@tailwind/fontFamily'

export const main = () => {
  document.body.style.backgroundColor = black
  document.body.style.fontFamily = sans
}
```

This is a lot better than trying to import the actual `tailwind.config.js` file to read these values as the values are fully resolved (merged with `defaultTheme`, plugins applied, etc.)

And it's a lot better than importing `tailwindcss/resolveConfig` and doing it in the app code because of the transitive dependencies taken on by that import.

It's also better than pre-compiling a static json file because these values are tree-shakeable. The entire generated json for the default tailwind config is ~100kb.  

The above example adds ~5kb to the overall bundle (and only because tailwind has so many default colors).

If you don't import from `@tailwindcss/*` nothing is added to the bundle (even if the imports are generated)

refers:

- none

## Type of change

**PATCH: backwards compatible change**



This PR includes breaking changes to the following core packages:

- none

This PR includes breaking changes to the follow extensions:

- none

## Dependencies

### Adds

- none

### Removes

- none