🩹 fix: [@roots/bud-criticalcss] critical import (#1763)
looks like critical has moved on to esm so there was an unresolved module error. we don't have unit tests for this extension yet (and it is marked in the bud docs as experimental so i don't really even feel bad). i wasn't aware until someone told me it was busted.

- adds a `try/catch` to the `critical.generate` function.
- uses namespace import since the default cjs export is gone.
- adds `base`, `src`, `html`, `width`, `height` functions (simple wrappers for `setOptions()`). 
- updates docs

refers:

- none

## Type of change

**PATCH: backwards compatible change**



This PR includes breaking changes to the following core packages:

- none

This PR includes breaking changes to the follow extensions:

- none

## Dependencies

### Adds

- none

### Removes

- none