🩹 fix: unresolvable optional dependency causes error (#1767)
Didn't catch this until doing e2e tests because deps are always resolved in the context of the repo. This stops optional dependencies which are not available from causing a TypeError.

refers:

- none

## Type of change

**PATCH: backwards compatible change**



This PR includes breaking changes to the following core packages:

- none

This PR includes breaking changes to the follow extensions:

- none

## Dependencies

### Adds

- none

### Removes

- none