📕 docs(none): 6.5.4 notes (#1862)
## ✨ improve(patch): clean up stack traces (#1773)

There should be less duplication of emitted errors, more meaningful stack traces, and highlighting of important parts of the error message.

## ✨ improve(patch): improve env interpolations (#1856)

Envvars sourced from .env files in the path to the project are now interpolated using env-expand.

If you have an unescaped string that causes the expansion to fail the error will be caught and gracefully downgraded to the basic handler.

## 🎉 feat(none): `bud upgrade` command (#1855)

Run `bud upgrade` to upgrade your project bud dependencies to the latest available version.

## 🩹 fix(patch): reload browser on watched files add/unlink (#1781)

When files are added or removed from a watched path, the browser will now reload.

## 🧪 test(none): migrate jest to vitest (#1850)

This should have no bearing on your project but we switched from jest to vitest and it's pretty cool.

## ℹ️ Release information

For more information [review the diff to see what's changed](https://github.com/roots/bud/compare/v6.5.3...v6.5.4).


refers:

- none

## Type of change

**NONE: internal change**



This PR includes breaking changes to the following core packages:

- none

This PR includes breaking changes to the follow extensions:

- none

## Dependencies

### Adds

- none

### Removes

- none