🩹 fix(patch): #1890 (#1891)
At first I thought I wanted to enforce not mixing/matching styles of import and would just update documentation. But, on second thought I decided it's better to stick to the current API -- the cost of maintaining it is essentially just one reducer.

- Fix: #1890
- Improve unit test: `sources/@roots/bud-api/src/methods/entry/index.test.ts`
- Add reproduction: `tests/reproductions/issue-1890`
- Add reproduction test: `tests/reproductions/issue-1890.test.ts` 

refers:

- #1890 

## Type of change

**PATCH: backwards compatible change**



This PR includes breaking changes to the following core packages:

- none

This PR includes breaking changes to the follow extensions:

- none

## Dependencies

### Adds

- none

### Removes

- none