🩹 fix(@roots/sage): serve options including hostname (#1983)
Fixes an issue where calling `bud.serve` with a full hostname (fully qualified `URL` or `string`; rather than an interface or a port) would cause hot assets to 500. this is due to an acorn implementation detail that is no longer needed, but it is easy to sidestep from the bud.js end by just removing the proxy request handler entirely and letting http-party do its thing.

refers:

- none

## Type of change

**PATCH: backwards compatible change**



This PR includes breaking changes to the following core packages:

- none

This PR includes breaking changes to the follow extensions:

- none

## Dependencies

### Adds

- none

### Removes

- none