✨ improve(@roots/bud): static config parsing (#2097)
More flexible static config parsing

https://github.com/roots/bud/blob/fdb48d5908621e8480999a201fded752d4b77c79/examples/config-yml/bud.config.yml#L1-L64

## Breaking change

Function args in a static config should be expressed as an array of arguments.

Example of code broken by this change:

```yml
assets:
  - image1.png
  - image2.png
```

How it should be modified:

```yml
assets:
  - ['image1.png', 'image2.png']
```

## Explanation

As a general rule: arguments passed to functions in a static config should be expressed as an array of arguments. It is always OK to do it this way. If you're unsure, just fall back on this rule.

Here's when it's OK to break this rule: 

If you are passing a single argument, **and the argument's value is not an array**, you can pass it as you would have prior to this change.

Examples: 

These function calls in a dynamic config:

```js
bud.entry('app.js')
```

```js
bud.entry({
  app: ['app.js']
})
```

can be expressed without an args array:

```yml
entry: 'app.js'
```

```yml
entry:
  app: 
    - app.js
```

But in this call we are passing an array value.

```js
entry(['app.js', 'app.css'])
```

And so we must pass it in an args array:

```yml
entry:
  - ['app.js', 'app.css']
```

Otherwise it will be interpreted as:

```js
bud.entry('app.js', 'app.css')
```

Which is not what we intended to express.

## Why

To better handle both:

- variadic functions
- functions which accept an array

refers:

- none

## Type of change

**MAJOR: breaking change**

This PR includes breaking changes to the following core packages:

- none

This PR includes breaking changes to the follow extensions:

- none

## Dependencies

### Adds

- none

### Removes

- none