✨ improve(@roots/wordpress-hmr): add loader (#2137)
- Adds loader for `@roots/wordpress-hmr`
- Improves `@roots/wordpress-hmr` api
- Registers loader in `@roots/bud-preset-wordpress`

Allows you to super easily register blocks with hot reload support. Blocks are registered for production build by the same call, as well.

This PR also includes support for plugins, formats, styles, variations and filters. 

See the updated docs attached to the PR.

## Using without loader

Still possible by importing from `@roots/wordpress-hmr`, but you have to handle providing the callbacks expected by the API.

## The loader

The loader replaces the `roots.register.[type]([dir])` with registration imports and calls.

The loader is generic (can be used in any webpack project). Exported from `@roots/wordpress-hmr/loader`

## Fully disabling

The easiest way to not use this functionality is to not use it (do it as you would have otherwise). 

If that isn't enough, for whatever reason, it can be fully disabled like this:

```js
bud.wp.set(`hmr`, false)
```

## Disabling snackbar notification(s)

There is a snackbar message when the hot reload fires for the first time. There will probably be a couple more in the future (errors, warnings).

Disable snackbar notices:

```js
bud.wp.set(`notify`, false)
```

refers:

- none

## Type of change

**PATCH: backwards compatible change**



This PR includes breaking changes to the following core packages:

- none

This PR includes breaking changes to the follow extensions:

- none

## Dependencies

### Adds

- none

### Removes

- none