Fix <alpha-value> Tailwind placeholder
Replace the special <alpha-value> placeholder. Tailwind CSS uses this placeholder to inject the alpha value when using an opacity modifier. The placeholder has been replaced with the default value of 1. Currently, the color value is copied as is, leading to broken colors such as --wp-preset-color--primary: rgba(0 0 0 / <alpha-value>). This fix ensures that the value is replaced with --wp-preset-color--primary: rgba(0 0 0 / 1).