✨ improve(patch): add @roots/bud-support/jsx-runtime (#2426)
If users are using a version of React from before React's work on `react/jsx-runtime`, then `react/jsx-runtime` will be `undefined`. 

The bud CLI requires this runtime and so all commands will fail to execute in that situation (or other equivalent situations).

This change re-exports the `react/jsx-runtime` via `@roots/bud-support/jsx-runtime` and configures typescript to include our package in the built bud.js code, since it will always be defined regardless of the user's application concerns. 

## Type of change

**PATCH: backwards compatible change**