✨ improve(patch): upgradeable community extensions (#2448)
Allow for community extensions to be versioned independently of bud.js packages.

Extension authors can add a `bud.version` field to their `package.json`:

```json
{
  "bud": {"version": ">=6.1.0"}
}
```

If no versioning information is present or a match cannot be found, bud.js will try to install `[extension-name]@latest` when upgrading.

This upgrade logic is applied to any `dependency` or `devDependency` which starts with `bud-` or includes `/bud-` (so packages can be namespaced).

## Type of change

**PATCH: backwards compatible change**