docs: `--target` becomes `--filter` (#2460)
@mikespainhower on Discord pointed out that `--target` seems to have been removed.

It looks like to me that the same functionality is now provided by `--filter`. Have tested `--filter` on a multi-instance setup and it targets the instance with the matching `bud.context.label`.

This PR is a search-replace on the docs page, `--target` becomes `--filter`.

Still haven't got a working Bud-dev setup running, so these will need compiling too (sorry).

## Type of change

**NONE: internal change**