✨ improve(patch): deterministic browserslist configs (#2467)
the goal is always to have deterministic output from bud.js managed builds. if two developers check out a bud.js based project and both run the same command they should expect to see identical behavior. it is literally the first project goal listed in the README: "bud.js wants to be... Reliable, yielding consistent and predictable behaviors regardless of specified options." 

browserslist's handling of query updates flies in the face of this goal; depending on the global state of the caniuse db it is very possible for output to be non-deterministic; emitting warnings for one developer and not another, even when using the same node version, package manager, etc. 

Compounding this frustration is the fact that browserslist emits its update reminder as a `console.warn` message, which may very well cause CI processes to fail.

pnpm, as a contrasting example, prints similar update reminders. However, they do so using `console.log`. This would be better.

This change makes all of our `@roots/browserslist-config` exports fully deterministic. This means that the configurations are constructed using unchanging targets, as opposed to queries based on market share (`> 2%`, `not dead`) or versions (`last 2 versions`, `last 3 versions`). As a result, build targets are now static and not reliant on non-deterministic browserslist datasets. 

We also set `BROWSERSLIST_IGNORE_OLD_DATA` to `true` if it is `undefined`. This prevents the console warnings from being emitted. 

Lastly, because our browserslist targets are static, we no longer need to manage upgrading browserslist datasets during postinstall. 

## Type of change

**PATCH: backwards compatible change**