🩹 fix(patch): WordPress module reload failures (#2530)
This fixes #2529 and other associated problems with hot module reload in certain WordPress environments.

## The problem

If `SCRIPT_DEBUG` is not set and a script declares `wp-react-refresh-runtime` as a dependency, WordPress will _**silently** omit the entire enqueue request_.

What the... 🙈

Bedrock installs don't trigger this behavior because Bedrock defines it:

https://github.com/roots/bedrock/blob/master/config/environments/development.php#L14

## Solution

This change omits including `wp-react-refresh-runtime` in `entrypoints.json` and does not externalize `react-refresh/runtime.js` if `SCRIPT_DEBUG` isn't set in a .env file somewhere.

## Type of change

**PATCH: backwards compatible change**