🩹 fix(patch): windows bootstrap path resolution (#2636)
First error to fix in order to provide native support for Windows. We have to start here because it happens during bootstrapping, so not much of the application can even run.

A similar fix may also need to be applied to `@roots/bud-framework/module` and `@roots/fs`. Not sure.

## From vite

They are calling `path.posix.normalize`, which is the same as `normalize` from `node:path/posix`:

https://github.com/vitejs/vite/blob/33d3cca78607e14e4d0b13e7012d065a94d184b1/packages/vite/src/node/utils.ts#L62-L64

Do we also need to prepend the `/` in windows? Don't know yet. See:

https://github.com/vitejs/vite/blob/33d3cca78607e14e4d0b13e7012d065a94d184b1/packages/vite/src/node/utils.ts#L63C30-L63C33

## Type of change

**PATCH: backwards compatible change**